add function to create list

+++ libft/ft_lstnew.c
@@ -6,7 +6,7 @@
 /*   By: vahemere <vahemere@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/22 00:41:01 by vahemere          #+#    #+#             */
-/*   Updated: 2021/11/23 14:20:14 by vahemere         ###   ########.fr       */
+/*   Updated: 2021/12/19 16:00:47 by vahemere         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ t_list	*ft_lstnew(void *content)
 	new = malloc(sizeof(t_list));
 	if (!new)
 		return (NULL);
-	new->content = content;
+	new->content = (int)content;
 	new->next = NULL;
 	return (new);
 }
